Move auth and IPAM constants to entity domains
