+++ frontend/app/src/entities/authentication/constants.ts
@@ -1,0 +1,2 @@
+export const ACCESS_TOKEN_KEY = "access_token";
+export const REFRESH_TOKEN_KEY = "refresh_token";
