+++ frontend/app/src/pages/graphql/index.tsx
@@ -5,7 +5,6 @@ import { useAtomValue } from "jotai";
 import { useQueryState } from "nuqs";
 
 import { CONFIG } from "@/shared/config/config";
-import { ACCESS_TOKEN_KEY } from "@/shared/config/localStorage";
 import { QSP } from "@/shared/config/qsp";
 import { datetimeAtom } from "@/shared/stores/time.atom";
 
@@ -13,6 +12,8 @@ import { useCurrentBranch } from "@/entities/branches/ui/branches-provider";
 
 import "graphiql/style.css";
 import "@graphiql/plugin-explorer/style.css";
+
+import { ACCESS_TOKEN_KEY } from "@/entities/authentication/constants";
 
 const plugins = [HISTORY_PLUGIN, explorerPlugin()];
