+++ frontend/app/src/shared/api/graphql/graphqlClientApollo.tsx
@@ -13,8 +13,8 @@ import { toast } from "react-toastify";
 import { queryClient } from "@/shared/api/rest/client";
 import { ALERT_TYPES, Alert } from "@/shared/components/ui/alert";
 import { CONFIG } from "@/shared/config/config";
-import { ACCESS_TOKEN_KEY } from "@/shared/config/localStorage";
 
+import { ACCESS_TOKEN_KEY } from "@/entities/authentication/constants";
 import { refreshAccessTokenQueryOptions } from "@/entities/authentication/domain/refresh-access-token.query";
 
 export const defaultOptions: DefaultOptions = {
