--- frontend/app/src/shared/config/localStorage.ts
@@ -1,2 +1,0 @@
-export const ACCESS_TOKEN_KEY = "access_token";
-export const IPAM_TREE_KEY = "ipam_tree_collapsed";
